+++ include/storage.h
@@ -3,7 +3,7 @@
 
 struct storage;
 
-struct storage *storage_init(int size, const char *path);
+struct storage *storage_init(unsigned long size, const char *path);
 
 int storage_read(const struct storage *dev, int offset, int size, void *buf);
 
@@ -13,6 +13,6 @@ int storage_destroy(struct storage *dev);
 
 void hex_dump(const void *data, int size);
 
-int get_size(const struct storage *dev);
+unsigned long get_size(const struct storage *dev);
 
 #endif
